baza danych pracuje

+++ app/models/Category.scala
@@ -1,8 +1,8 @@
 package models
 import play.api.libs.json._
+import play.api.libs.json._
 
-case class Category(id: Long, name: String)
-
+case class Category(id: Int, name: String)
 
 object Category {
   implicit val categoryFormat = Json.format[Category]
